Move insights export below analytics
